Add treetop.submit(elm) function and remove form component
The form component was patching the .submit function on the element
instance. This is not a good way to do it. If the user wishes to submit
a form without using submit event I have added a treetop.submit function
to allow them to do so. This is a *far* better solution.